requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-bcrypt==1.0.1
python-dotenv==1.0.0
pymysql==1.1.0
cryptography
b2sdk
gunicorn>=20.1.0